File example.py has been refactored to menu
